refactor(exec): Move internal window files under exec/window
The window operator code mixes public API with internal implementation
details in velox/exec/. Window, WindowFunction, and WindowPartition are
the public API used by window function implementations; the build
strategies, aggregate-as-window adapter, and frame/peer-group helpers are
internal to the operator.

Move the internal pieces to velox/exec/window/ under the
facebook::velox::exec::window namespace, leaving the public API
(Window, WindowFunction, WindowPartition) unchanged in velox/exec/ and
the facebook::velox::exec namespace. No files outside velox/exec/ change.
