+++ velox/exec/window/CMakeLists.txt
@@ -1,0 +1,35 @@
+# Copyright (c) Facebook, Inc. and its affiliates.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+velox_add_library(
+  velox_exec_window
+  AggregateWindow.cpp
+  PartitionStreamingWindowBuild.cpp
+  RowsStreamingWindowBuild.cpp
+  SortWindowBuild.cpp
+  SubPartitionedSortWindowBuild.cpp
+  WindowBuild.cpp
+  HEADERS
+  AggregateWindow.h
+  KRangeFrameBound.h
+  PartitionStreamingWindowBuild.h
+  PeerGroupComputation.h
+  RowsStreamingWindowBuild.h
+  SortWindowBuild.h
+  SubPartitionedSortWindowBuild.h
+  WindowBuild.h
+  WindowPartitionAccessor.h
+)
+
+velox_link_libraries(velox_exec_window velox_exec velox_vector)
