+++ velox/exec/window/WindowBuild.h
@@ -18,7 +18,7 @@
 #include "velox/exec/RowContainer.h"
 #include "velox/exec/WindowPartition.h"
 
-namespace facebook::velox::exec {
+namespace facebook::velox::exec::window {
 
 /// The Window operator needs to see all input rows, and separate them into
 /// partitions based on a partitioning key. There are many approaches to do
@@ -136,4 +136,4 @@ class WindowBuild {
   vector_size_t numRowsPerOutput_;
 };
 
-} // namespace facebook::velox::exec
+} // namespace facebook::velox::exec::window
